Add transportation cost to objective
The flows, as currently written, only depend on the network edges and the hour. They are independent of the simulation year since each edge is already associated with a specific year. This made adding transportation cost more difficult, since the discount factor depends on the year. To overcome this challenge, I made the transportation cost coefficent also depend on the simulation year. If the simulation year does not match the year of an edge, then the coefficient is set to zero to nullify the discount factor.