Document model
I added model documentation using Julia's Documenter package. The documentation could be published to a GitHub Pages website. We could currently not do this, however, because GitHub Pages is only free for public code.

I added Doc Strings to all key functions in the model except the ones which process the results. Furthermore, I created a new folder 'docs' which contains the markdown files for the code documentation. The HTML files for GitHub Pages are not pushed, but can be generated by running the script "make.jl". Upon running the script, the HTML files will be added to a new directory "build" in the "docs" folder.